Add template for writing Spanner Change Stream records into GCS.
PiperOrigin-RevId: 440217254

+++ v2/googlecloud-to-googlecloud/src/main/java/com/google/cloud/teleport/v2/transforms/WriteDataChangeRecordsToGcsAvro.java
@@ -103,7 +103,7 @@ public abstract class WriteDataChangeRecordsToGcsAvro
     @Description(
         "The maximum number of output shards produced when writing. Default number is runner"
             + " defined.")
-    @Default.Integer(1)
+    @Default.Integer(20)
     Integer getNumShards();
 
     void setNumShards(Integer numShards);
