+++ v2/googlecloud-to-googlecloud/src/main/java/com/google/cloud/teleport/v2/transforms/WriteDataChangeRecordsToGcsText.java
@@ -121,7 +121,7 @@ public abstract class WriteDataChangeRecordsToGcsText
     @Description(
         "The maximum number of output shards produced when writing. Default number is runner"
             + " defined.")
-    @Default.Integer(1)
+    @Default.Integer(20)
     Integer getNumShards();
 
     void setNumShards(Integer numShards);
